pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "catalog2md"
version = "1.0.0"
description = "Convert technical PDF catalogs to RAG-optimized Markdown files"
requires-python = ">=3.10"
dependencies = [
    "pdfplumber>=0.10.0",
    "anthropic>=0.30.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pypdf>=4.0",
    "pdf2image>=1.17.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
docling = [
    "docling>=2.9.0",
]

[project.scripts]
catalog2md = "catalog2md.cli:main"
